+++ Engine/Source/Developer/Mac/MacTargetPlatform/Private/LocalMacTargetDevice.h
@@ -188,7 +188,7 @@ public:
 		return false;
 	}
 
-	virtual bool TerminateProcess( const int32 ProcessId ) override
+	virtual bool TerminateProcess( const int64 ProcessId ) override
 	{
 		return false;
 	}
